ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
email-validator>=2.0.0
Pillow==10.4.0
numpy>=1.21.0
easyocr==1.7.1
python-multipart==0.0.9
